requirements.txt
certifi==2019.11.28
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
django-filter==2.2.0
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==20.0.4
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
wincertstore==0.2
